AI_backend/requirement.txt
langchain_core
python-dotenv
langchain-huggingface
flask